Secuencias funcionando para Sql Server Proyecto Clientes
Se completa la funcionalidad del uso de secuencias para el proyecto de
Clientes con Sql Server y Oracle.

+++ ProyectosNetbeansWS/TouresBalonClientesWS/src/main/java/aes/pica/touresbalon/touresbalonproductosws/entidades/productos/TarifaHospedaje.java
@@ -1,7 +1,8 @@
 package aes.pica.touresbalon.touresbalonproductosws.entidades.productos;
-// Generated 21-sep-2016 23:29:49 by Hibernate Tools 4.3.1
+// Generated 24/09/2016 07:45:59 PM by Hibernate Tools 4.3.1
 
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -13,7 +14,7 @@ public class TarifaHospedaje  implements java.io.Serializable {
 
      private int idHospedaje;
      private String nombreHospedaje;
-     private Long precio;
+     private BigDecimal precio;
      private Set productos = new HashSet(0);
 
     public TarifaHospedaje() {
@@ -23,7 +24,7 @@ public class TarifaHospedaje  implements java.io.Serializable {
     public TarifaHospedaje(int idHospedaje) {
         this.idHospedaje = idHospedaje;
     }
-    public TarifaHospedaje(int idHospedaje, String nombreHospedaje, Long precio, Set productos) {
+    public TarifaHospedaje(int idHospedaje, String nombreHospedaje, BigDecimal precio, Set productos) {
        this.idHospedaje = idHospedaje;
        this.nombreHospedaje = nombreHospedaje;
        this.precio = precio;
@@ -44,11 +45,11 @@ public class TarifaHospedaje  implements java.io.Serializable {
     public void setNombreHospedaje(String nombreHospedaje) {
         this.nombreHospedaje = nombreHospedaje;
     }
-    public Long getPrecio() {
+    public BigDecimal getPrecio() {
         return this.precio;
     }
     
-    public void setPrecio(Long precio) {
+    public void setPrecio(BigDecimal precio) {
         this.precio = precio;
     }
     public Set getProductos() {
